[ALPS07233972] Disp: Fix coverity issue
typecast to prevent coverity issue

MTK-Commit-Id: f300ed9003214933504ca6e78ff01181e316c685

CR-Id: ALPS07233972
Feature: display_mobile

+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_dsi.c
@@ -6767,7 +6767,7 @@ unsigned long long mtk_dsi_get_frame_hrt_bw_base_by_mode(
 		bw_base = bw_base * mode->vtotal / mode->vdisplay;
 		bw_base = bw_base / 1000;
 	} else {
-		bw_base = data_rate * dsi->lanes * compress_rate * 4;
+		bw_base = (unsigned long long) data_rate * dsi->lanes * compress_rate * 4;
 		bw_base = bw_base / bpp / 100;
 	}
